chore(suite-native): Mobile Trade: Refactor general/quotesUtils to buy/quotesUtils
