+++ suite-native/module-trading/src/utils/buy/quotesUtils.ts
